\name{CIBMTR_Params}
\alias{CIBMTR_Params}
\docType{data}
\title{
Estimates for model parameters from semi-competing risks analysis of the CIBMTR data using Weibull illness-death model.
}
\description{
Estimates for model parameters from semi-competing risks analysis of the CIBMTR data using Weibull illness-death model.
}
\usage{data("CIBMTR_Params")}
\format{
  The format is a list of 10 components corresponding to parameters for Weibull illness-death model.
}
\seealso{
	\code{\link{CIBMTR}}
}
\examples{
data(CIBMTR_Params)
}
\keyword{datasets}
