\name{old.functions}
\alias{BayesID}
\alias{BayesSurv}
\alias{FreqID}
\alias{FreqSurv}
\alias{initiate.startValues}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Old functions
}
\description{
Since version 2.5, the functions BayesID(), BayesSurv(), FreqID(), FreqSurv(), initiate.startValues() have been renamed as BayesID_HReg(), BayesSurv_HReg(), FreqID_HReg(), FreqSurv_HReg(), initiate.startValues_HReg(), respectively. If one of the old functions is called, a warning message will be displayed with the corresponding new function name.

}
\usage{
BayesID(...)
BayesSurv(...)
FreqID(...)
FreqSurv(...)
initiate.startValues(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
	\item{...}{
    arguments used for the old functions.
}
}


